# api/openapi.v2.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v2/auth/register": {
      "post": {
        "operationId": "Auth_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "meta"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Registration"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {},
                      "maxProperties": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The input is invalid; `errors` names each field at fault. Codes: `bad_request`, `invalid_fields`, `invalid_id`, `user_already_active`, `cannot_target_self`, `last_active_admin`, `credit_fields_on_non_credit`, `no_files`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state (a duplicate, a resource in use). Codes: `conflict`, `duplicate`, `constraint_violation`, `account_has_transactions`, `category_has_children`, `reassignment_required`, `reassignment_target_inside`, `template_requires_empty`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Well-formed but breaks a business rule. Codes: `validation_failed`, `check_violation`, `weak_password`, `password_reset_not_managed`, `splits_unbalanced`, `amount_breaks_splits`, `transfer_same_account`, `transfer_leg_locked`, `account_not_owned`, `category_not_owned`, `category_cycle`, `category_too_deep`, `multi_payment_requires_concept`, `multi_payment_requires_recurring`, `multi_payment_excludes_auto_paid`, `merge_into_itself`, `merge_requires_concepts`, `merge_source_has_children`, `interpretation_needs_text`, `concept_archived`, `cost_center_cannot_classify`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached; retry later. Codes: `rate_limited`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. Codes: `internal_error`, `identity_provider_failed`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "tags": ["AuthV2"]
      }
    },
    "/api/v2/auth/login": {
      "post": {
        "operationId": "Auth_login",
        "parameters": [
          {
            "name": "x-coco-client",
            "in": "header",
            "description": "Native clients send `native`: the refresh token then travels in the body (`refreshToken`) instead of the httpOnly `coco_refresh` cookie the web uses.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["native"]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "meta"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Session"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {},
                      "maxProperties": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The input is invalid; `errors` names each field at fault. Codes: `bad_request`, `invalid_fields`, `invalid_id`, `user_already_active`, `cannot_target_self`, `last_active_admin`, `credit_fields_on_non_credit`, `no_files`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token (or session to renew). Codes: `unauthenticated`, `invalid_credentials`, `invalid_token`, `session_expired`, `session_revoked`, `wrong_current_password`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The user may not do this. Codes: `forbidden`, `account_not_enabled`, `account_pending_approval`, `account_suspended`, `credentials_not_managed`, `real_accounts_protected`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached; retry later. Codes: `rate_limited`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. Codes: `internal_error`, `identity_provider_failed`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "tags": ["AuthV2"]
      }
    },
    "/api/v2/auth/refresh": {
      "post": {
        "operationId": "Auth_refresh",
        "parameters": [
          {
            "name": "x-coco-client",
            "in": "header",
            "description": "Native clients send `native`: the refresh token then travels in the body (`refreshToken`) instead of the httpOnly `coco_refresh` cookie the web uses.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["native"]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "meta"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Session"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {},
                      "maxProperties": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The input is invalid; `errors` names each field at fault. Codes: `bad_request`, `invalid_fields`, `invalid_id`, `user_already_active`, `cannot_target_self`, `last_active_admin`, `credit_fields_on_non_credit`, `no_files`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token (or session to renew). Codes: `unauthenticated`, `invalid_credentials`, `invalid_token`, `session_expired`, `session_revoked`, `wrong_current_password`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The user may not do this. Codes: `forbidden`, `account_not_enabled`, `account_pending_approval`, `account_suspended`, `credentials_not_managed`, `real_accounts_protected`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached; retry later. Codes: `rate_limited`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. Codes: `internal_error`, `identity_provider_failed`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "tags": ["AuthV2"]
      }
    },
    "/api/v2/auth/logout": {
      "post": {
        "operationId": "Auth_logout",
        "parameters": [
          {
            "name": "x-coco-client",
            "in": "header",
            "description": "Native clients send `native`: the refresh token then travels in the body (`refreshToken`) instead of the httpOnly `coco_refresh` cookie the web uses.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["native"]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Done. No body."
          },
          "400": {
            "description": "The input is invalid; `errors` names each field at fault. Codes: `bad_request`, `invalid_fields`, `invalid_id`, `user_already_active`, `cannot_target_self`, `last_active_admin`, `credit_fields_on_non_credit`, `no_files`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached; retry later. Codes: `rate_limited`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. Codes: `internal_error`, `identity_provider_failed`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "tags": ["AuthV2"]
      }
    },
    "/api/v2/auth/logout-all": {
      "post": {
        "operationId": "Auth_logoutAll",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Done. No body."
          },
          "401": {
            "description": "No valid access token (or session to renew). Codes: `unauthenticated`, `invalid_credentials`, `invalid_token`, `session_expired`, `session_revoked`, `wrong_current_password`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached; retry later. Codes: `rate_limited`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. Codes: `internal_error`, `identity_provider_failed`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["AuthV2"]
      }
    },
    "/api/v2/auth/me": {
      "get": {
        "operationId": "Auth_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "meta"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Me"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {},
                      "maxProperties": 0
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No valid access token (or session to renew). Codes: `unauthenticated`, `invalid_credentials`, `invalid_token`, `session_expired`, `session_revoked`, `wrong_current_password`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached; retry later. Codes: `rate_limited`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. Codes: `internal_error`, `identity_provider_failed`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["AuthV2"]
      }
    },
    "/api/v2/auth/change-password": {
      "post": {
        "operationId": "Auth_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Done. No body."
          },
          "400": {
            "description": "The input is invalid; `errors` names each field at fault. Codes: `bad_request`, `invalid_fields`, `invalid_id`, `user_already_active`, `cannot_target_self`, `last_active_admin`, `credit_fields_on_non_credit`, `no_files`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token (or session to renew). Codes: `unauthenticated`, `invalid_credentials`, `invalid_token`, `session_expired`, `session_revoked`, `wrong_current_password`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Well-formed but breaks a business rule. Codes: `validation_failed`, `check_violation`, `weak_password`, `password_reset_not_managed`, `splits_unbalanced`, `amount_breaks_splits`, `transfer_same_account`, `transfer_leg_locked`, `account_not_owned`, `category_not_owned`, `category_cycle`, `category_too_deep`, `multi_payment_requires_concept`, `multi_payment_requires_recurring`, `multi_payment_excludes_auto_paid`, `merge_into_itself`, `merge_requires_concepts`, `merge_source_has_children`, `interpretation_needs_text`, `concept_archived`, `cost_center_cannot_classify`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached; retry later. Codes: `rate_limited`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure. Codes: `internal_error`, `identity_provider_failed`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["AuthV2"]
      }
    },
...
  },
  "info": {
    "title": "Coco API",
    "description": "Contract v2, served under `/api/v2`: the v1 API in English and camelCase, with every\nlist paginated as `{ data, meta: { page, perPage, total } }`.\n\nEvery successful JSON response is wrapped as `{ data, meta }`. Every error is\n`application/problem+json` (RFC 9457): `{ type, title, status, detail, code, errors? }`.\nSwitch on `code`: stable, in English, one per business rule. `errors` lists the fields\nthat failed.\nMessages meant for the user are in Spanish.\n\nIdentifiers are integers (int64) and money travels as a decimal string.",
    "version": "2",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 255,
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128
          },
          "displayName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 255
          }
        },
        "required": ["email", "password", "displayName"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 255,
            "format": "email"
          },
          "password": {
            "type": "string",
            "maxLength": 128
          }
        },
        "required": ["email", "password"]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "maxLength": 128
          },
          "newPassword": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": ["currentPassword", "newPassword"]
      },
      "ProblemFieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Path of the field at fault (`splits.0.amount`), when there is one."
          },
          "message": {
            "type": "string",
            "description": "In Spanish: the client may show it next to the field."
          }
        },
        "required": ["message"]
      },
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "description": "Identifies the kind of problem: `https://dev-cocoapp.viteri.me/problems/<code>`.",
            "example": "https://dev-cocoapp.viteri.me/problems/not_found"
          },
          "code": {
            "enum": [
              "bad_request",
              "invalid_fields",
              "invalid_id",
              "unauthenticated",
              "forbidden",
              "not_found",
              "conflict",
              "duplicate",
              "constraint_violation",
              "payload_too_large",
              "unsupported_media_type",
              "validation_failed",
              "check_violation",
              "rate_limited",
              "internal_error",
              "service_unavailable",
              "invalid_credentials",
              "invalid_token",
              "session_expired",
              "session_revoked",
              "wrong_current_password",
              "account_not_enabled",
              "account_pending_approval",
              "account_suspended",
              "credentials_not_managed",
              "real_accounts_protected",
              "weak_password",
              "identity_provider_failed",
              "user_already_active",
              "cannot_target_self",
              "last_active_admin",
              "password_reset_not_managed",
              "splits_unbalanced",
              "amount_breaks_splits",
              "transfer_same_account",
              "transfer_leg_locked",
              "account_not_owned",
              "category_not_owned",
              "account_has_transactions",
              "credit_fields_on_non_credit",
              "category_cycle",
              "category_too_deep",
              "category_has_children",
              "reassignment_required",
              "reassignment_target_inside",
              "template_requires_empty",
              "multi_payment_requires_concept",
              "multi_payment_requires_recurring",
              "multi_payment_excludes_auto_paid",
              "merge_into_itself",
              "merge_requires_concepts",
              "merge_source_has_children",
              "interpretation_needs_text",
              "concept_archived",
              "cost_center_cannot_classify",
              "no_files",
              "file_type_not_allowed",
              "file_content_mismatch",
              "file_too_large",
              "image_format_unsupported",
              "image_processing_unavailable",
              "receipt_file_missing",
              "database_unavailable"
            ],
            "type": "string",
            "description": "Stable, machine-readable: one per business rule (`problem-codes.ts`).",
            "example": "not_found"
          },
          "title": {
            "type": "string",
            "description": "Short summary of the kind of problem, in Spanish; the same for every occurrence."
          },
          "status": {
            "type": "number",
            "description": "The HTTP status, repeated."
          },
          "detail": {
            "type": "string",
            "description": "What happened this time, in Spanish: the client may show it as is."
          },
          "errors": {
            "description": "One per invalid field; present only when the input itself is at fault.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemFieldError"
            }
          }
        },
        "required": ["type", "code", "title", "status", "detail"]
      },
      "Registration": {
        "type": "object",
        "properties": {
          "pendingApproval": {
            "type": "boolean",
            "description": "Whether an administrator must approve the account before it can sign in."
          },
          "message": {
            "type": "string",
            "description": "In Spanish: the client shows it as is."
          }
        },
        "required": ["pendingApproval", "message"]
      },
...
      "Session": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "number",
            "description": "Seconds until `accessToken` expires."
          },
          "user": {
            "$ref": "#/components/schemas/Profile"
          },
          "refreshToken": {
            "type": "string",
            "description": "Native clients only (`X-Coco-Client: native`); the web gets an httpOnly cookie instead."
          }
        },
        "required": ["accessToken", "expiresIn", "user"]
      },
      "RefreshInput": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "maxLength": 4096
          }
        }
      },
      "Me": {
        "type": "object",
        "properties": {
          "role": {
            "enum": ["admin", "user"],
            "type": "string"
          },
          "status": {
            "enum": ["pending", "active", "suspended"],
            "type": "string"
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["flags_canary"]
            }
          },
          "id": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["role", "status", "features", "id", "email", "displayName", "createdAt"]
      },
...
    }
  }
}